CSharp -- Restructure and review documentation code examples
