+++ Dash.NET.Giraffe.CSharp.Tests/Template_Example.cs
@@ -1,0 +1,22 @@
+﻿using System;
+using System.Collections.Generic;
+using System.Linq;
+using System.Text;
+using System.Threading.Tasks;
+using Dash.NET.CSharp.Giraffe;
+using Dash.NET.CSharp.DCC;
+using static Dash.NET.CSharp.Dsl;
+using Plotly.NET;
+using Microsoft.Extensions.Logging;
+using Giraffe;
+
+namespace Documentation.Examples
+{
+    class Page_Section
+    {
+        public static void RunExample()
+        {
+
+        }
+    }
+}
